test: update timeseries extentions case desc
