+++ test/cases/24-TimeSeriesExtensions/03-TimeWindow/ans/interval.csv
@@ -1372,57 +1372,9 @@ taos> select _wstart, _wend, _wduration, count(*) from test.st where (ts >= '202
  2020-12-07 00:00:00.000 | 2021-01-07 00:00:00.000 |            2678400000 |                     1 |
 
 taos> select * from test.sta;
-         _wstart         |          _wend          |      _wduration       |       count(*)        |       group_id        |
-============================================================================================================================
- 2020-10-01 00:00:00.000 | 2020-10-01 00:01:00.000 |                 60000 |                     1 |                     0 |
- 2020-10-01 00:01:00.000 | 2020-10-01 00:02:00.000 |                 60000 |                     1 |                     0 |
- 2020-10-01 00:02:00.000 | 2020-10-01 00:03:00.000 |                 60000 |                     1 |                     0 |
- 2020-10-01 00:03:00.000 | 2020-10-01 00:04:00.000 |                 60000 |                     1 |                     0 |
- 2020-10-01 00:04:00.000 | 2020-10-01 00:05:00.000 |                 60000 |                     1 |                     0 |
- 2020-10-01 00:05:00.000 | 2020-10-01 00:06:00.000 |                 60000 |                     1 |                     0 |
- 2020-10-01 00:06:00.000 | 2020-10-01 00:07:00.000 |                 60000 |                     1 |                     0 |
- 2020-10-01 00:07:00.000 | 2020-10-01 00:08:00.000 |                 60000 |                     1 |                     0 |
-
 taos> select * from test.stb;
-         _wstart         |          _wend          |      _wduration       |       count(*)        |       group_id        |
-============================================================================================================================
- 2020-11-01 22:51:00.000 | 2020-11-01 23:51:00.000 |               3600000 |                     1 |                     0 |
- 2020-11-01 23:18:00.000 | 2020-11-02 00:18:00.000 |               3600000 |                     1 |                     0 |
- 2020-11-01 23:45:00.000 | 2020-11-02 00:45:00.000 |               3600000 |                     1 |                     0 |
-
 taos> select * from test.stc;
-         _wstart         |          _wend          |      _wduration       |       count(*)        |       group_id        |
-============================================================================================================================
- 2020-10-17 23:32:00.000 | 2020-11-17 23:32:00.000 |            2678400000 |                     1 |                     0 |
- 2020-10-30 23:32:00.000 | 2020-11-30 23:32:00.000 |            2678400000 |                     1 |                     0 |
- 2020-11-12 23:32:00.000 | 2020-12-12 23:32:00.000 |            2592000000 |                     1 |                     0 |
-
 taos> select * from test.std;
-         _wstart         |          _wend          |      _wduration       |       count(*)        |       group_id        |
-============================================================================================================================
- 2020-10-09 01:23:00.000 | 2020-10-09 01:23:01.000 |                  1000 |                     1 |                     0 |
- 2020-11-09 01:23:00.000 | 2020-11-09 01:23:01.000 |                  1000 |                     1 |                     0 |
- 2020-12-09 01:23:00.000 | 2020-12-09 01:23:01.000 |                  1000 |                     1 |                     0 |
-
 taos> select * from test.ste;
-         _wstart         |          _wend          |      _wduration       |       count(*)        |       group_id        |
-============================================================================================================================
- 2020-12-08 08:23:00.001 | 2020-12-09 08:23:00.001 |              86400000 |                   420 |                     0 |
- 2020-12-09 01:23:00.001 | 2020-12-10 01:23:00.001 |              86400000 |                   556 |                     0 |
-
 taos> select * from test.stf;
-         _wstart         |          _wend          |      _wduration       |       count(*)        |       group_id        |
-============================================================================================================================
- 2020-10-09 01:23:00.000 | 2020-11-09 01:23:00.000 |            2678400000 |                 44640 |                     0 |
- 2020-11-09 01:23:00.000 | 2020-12-09 01:23:00.000 |            2592000000 |                 43200 |                     0 |
- 2020-12-09 01:23:00.000 | 2021-01-09 01:23:00.000 |            2678400000 |                   557 |                     0 |
-
 taos> select * from test.stg;
-         _wstart         |          _wend          |      _wduration       |       count(*)        |       group_id        |
-============================================================================================================================
- 2020-10-14 01:23:00.000 | 2020-11-14 01:23:00.000 |            2678400000 |                  7200 |                     0 |
- 2020-10-27 01:23:00.000 | 2020-11-27 01:23:00.000 |            2678400000 |                 25920 |                     0 |
- 2020-11-09 01:23:00.000 | 2020-12-09 01:23:00.000 |            2592000000 |                 43200 |                     0 |
- 2020-11-22 01:23:00.000 | 2020-12-22 01:23:00.000 |            2592000000 |                 25037 |                     0 |
- 2020-12-05 01:23:00.000 | 2021-01-05 01:23:00.000 |            2678400000 |                  6317 |                     0 |
-
