+++ test/ci/cases.task
@@ -1,6 +1,5 @@
 #caseID,rerunTimes,Run with Sanitizer,casePath,caseCommand
 #NA,NA,y or n,.,./ci/pytest.sh pytest cases/01-DataTypes/test_datatype_bigint.py
-#examples
 
 # 01-DataTypes
 ,,y,.,./ci/pytest.sh pytest cases/01-DataTypes/test_datatype_blob.py
@@ -270,19 +269,12 @@
 ## 01-PartitionBy
 ,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/01-PartitionBy/test_partitonby.py
 ## 02-Fill
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/02-Fill/test_fill_force.py
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/02-Fill/test_fill_stb.py
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/02-Fill/test_fill_us.py
 ,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/02-Fill/test_fill.py
 ## 03-TimeWindow
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval_bug.py
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval_empty_range_scl.py
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval_empty_range.py
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval_function.py
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval_offset.py
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval.py
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval2.py
-,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_sliding.py
+,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval_auto.py
+,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval_basic.py
+,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval_bug_ts5400.py
+,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/03-TimeWindow/test_interval_timezone.py
 ## 04-StateWindow
 ## 05-EventWindow
 ,,y,.,./ci/pytest.sh pytest cases/24-TimeSeriesExtensions/05-EventWindow/test_event.py
